\name{fractalart-package}
\alias{fractalart-package}
\alias{fractalart}
\docType{package}
\title{fractalart
\packageTitle{fractalart}
}
\description{
\packageDescription{fractalart}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fractalart}
\packageIndices{fractalart}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fractalart}

Maintainer: \packageMaintainer{fractalart}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
